Redo file structure and routes file
This commit deletes the original file structure I had because I realised I hadn't nested it within a `posts` folder.
While the blengine is just for posts now, there may be a time when I want to add other things (like projects, or playlists, or even photos) and my previous unfoldered approach was not the most extensible.

I have also made a modification to the router.ex because I realised I had the `/posts/new` route pointing at the wrong file. It should be pointed at the form_live.ex file, as that is what I'll be using to generate new posts.