Create post and tag files, move liveview files
I used ecto to generate the "model" modules for posts and tags. This was helpful as it autogenerate the changeset as well as the schema structure, which I missed when I thought I had to create them manually.
The downside of this approach was that I had already manually generated migrations to create the tags and posts tables. There is therefore a duplication in the migrations, but when I ran them I was just returned an error stating that those tables already exist. I'm hoping this won't cause problems later, but only time will tell...

This commit also includes the newly move liveview files so that they live in their "blog_posts" folder. I might have to rename this to "posts" for consistency, but as far as I can tell it's not causing any issues thus far.